#231 refactor : 알람 로직 분리
